+++ src/Firebase/ServiceAccount.php
@@ -16,6 +16,16 @@ class ServiceAccount
     private $clientId;
     private $clientEmail;
     private $privateKey;
+    /** @var string|null */
+    private $filePath;
+
+    /**
+     * @return string|null
+     */
+    public function getFilePath()
+    {
+        return $this->filePath;
+    }
 
     public function getProjectId(): string
     {
@@ -165,10 +175,14 @@ class ServiceAccount
         }
 
         try {
-            return self::fromJson($json);
+            $serviceAccount = self::fromJson($json);
         } catch (Throwable $e) {
             throw new InvalidArgumentException(\sprintf('%s could not be parsed to a Service Account: %s', $filePath, $e->getMessage()));
         }
+
+        $serviceAccount->filePath = $filePath;
+
+        return $serviceAccount;
     }
 
     public static function withProjectIdAndServiceAccountId(string $projectId, string $serviceAccountId): self